requirements.txt
fastapi==0.75.0
psycopg2-binary==2.9.3
uvicorn==0.17.6
python_dotenv==0.19.2
pyyaml==6.0
